requirements.txt
streamlit>=1.30
pandas>=2.2
requests>=2.31
python-dotenv>=1.0
